Clean up a couple of warnings.

+++ Branch-TestBed/Branch-TestBed/ExampleDeepLinkingController.m
@@ -29,7 +29,7 @@
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
     
     // Create 16 byte MD5 hash value, store in buffer
-    CC_MD5(ptr, strlen(ptr), md5Buffer);
+    CC_MD5(ptr, (unsigned int)strlen(ptr), md5Buffer);
     
     // Convert MD5 value in the buffer to NSString of hex values
     NSMutableString *emailHash = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
